[Feature] rebuild the conversation flow and code for loan record retrieval and loan extension
1. build the loan record and loan extension conversation flow and code according to the user requirement as far as possible

pending items in this version:
a. loan record is independent from extend loan and the two are not yet linked as user requirement doc
b. loan record retrieval is based on the past 6 months, not the current sem ending
c. loan record not displaying author information yet (pending api side changes by long ke to include a change of search end point)
d. not yet test the info display with overdue book items. need to amend the data to get this tested.